pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "juliuscaesar"
version = "0.1.0"
description = "Local-first personal assistant runtime with gateway, watchdog, heartbeat, memory, and voice tooling."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "python-dotenv",
  "pyyaml",
  "requests",
  "websocket-client",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.6",
]
voice = [
  "dashscope",
]
slack = [
  "slack-bolt",
]
discord = [
  "discord.py",
]

[tool.setuptools]
package-dir = { "" = "lib" }
py-modules = ["jc_paths"]

[tool.setuptools.packages.find]
where = ["lib"]

[tool.pytest.ini_options]
testpaths = [
  "tests",
  "tests/gateway",
  "tests/gateway/recovery",
  "tests/watchdog",
]
python_files = ["test_*.py"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["F821"]

[tool.coverage.run]
source = ["lib"]
branch = true
